Tier@number slot changed from 'numericr' to 'integer'

+++ R/Tier-class.R
@@ -17,5 +17,5 @@ setClass(
   Class = 'Tier',
   contains = c(),
   slots    = c(name   = 'character',
-               number = 'numeric')
+               number = 'integer')
 )
